Added the ability to add select elements to the toolbar instead of buttons with triggers
