+++ src/core/constants.ts
@@ -11,7 +11,7 @@
 import type { HTMLTagNames, IDictionary } from 'jodit/types';
 
 export const APP_VERSION = process.env.APP_VERSION as string;
-// eslint-disable-next-line prettier/prettier
+// prettier-ignore
 export const ES: 'es5' | 'es2015' | 'es2018' | 'es2021' = process.env.TARGET_ES as 'es2015';
 export const IS_ES_MODERN = process.env.IS_ES_MODERN as unknown as boolean;
 export const IS_ES_NEXT = process.env.IS_ES_NEXT as unknown as boolean;
