+++ src/core/helpers/utils/css.ts
@@ -45,6 +45,7 @@ export function css(
  * Get the value of a computed style property for the first element in the set of matched elements or set one or
  * more CSS properties for every matched element
  *
+ * @param element - HTML element
  * @param key - An object of property-value pairs to set. A CSS property name.
  * @param value - A value to set for the property.
  * @param onlyStyleMode - Get value from style attribute, without calculating
