+++ src/core/ui/helpers/get-control-type.ts
@@ -15,7 +15,7 @@ import type {
 	IControlTypeStrong,
 	IDictionary
 } from 'jodit/types';
-import { ConfigFlatten, isString } from 'jodit/core/helpers';
+import { ConfigFlatten, isArray, isString } from 'jodit/core/helpers';
 import { Config } from 'jodit/config';
 
 /**
@@ -72,10 +72,24 @@ export function findControlType(
 		key = namespaceOrKey;
 	}
 
+	const list = store[key]?.list;
+
 	return store[key]
 		? {
 				name: key,
-				...ConfigFlatten(store[key])
+				...ConfigFlatten(store[key]),
+				list: isArray(list)
+					? (<Array<string>>list).reduce(
+							(
+								acc: IDictionary<string | number>,
+								k: string | number
+							) => {
+								acc[k] = k;
+								return acc;
+							},
+							{}
+					  )
+					: list
 		  }
 		: undefined;
 }
