+++ src/modules/file-browser/config.ts
@@ -405,7 +405,7 @@ Config.prototype.controls.filebrowser = {
 
 	filter: {
 		isInput: true,
-		getContent: (filebrowser: IFileBrowser, _, b): HTMLElement => {
+		getContent: (filebrowser: IFileBrowser, b): HTMLElement => {
 			const oldInput = b.container.querySelector('.jodit-input');
 
 			if (oldInput) {
