+++ src/modules/toolbar/collection/collection.less
@@ -68,6 +68,7 @@
 	.jodit-ui-list();
 
 	.jodit-toolbar-content,
+	.jodit-toolbar-select,
 	.jodit-toolbar-button {
 		padding: 0;
 		margin: var(--margin-v) 1px;
