+++ src/plugins/table/config.ts
@@ -43,7 +43,9 @@ Config.prototype.controls.table = {
 			'table table-dark': 'Bootstrap Dark'
 		}
 	},
-	popup: (editor: IJodit, current, control, close, button) => {
+	popup: (editor: IJodit, current, close, button) => {
+		const control = button.control;
+
 		const default_rows_count =
 				control.data && control.data.rows ? control.data.rows : 10,
 			default_cols_count =
