+++ src/styles/mixins.less
@@ -18,7 +18,9 @@
 
 .active() {
 	background-image: none;
-	box-shadow: inset 0 2px 4px rgb(0 0 0 / 30%), 0 1px 2px rgb(0 0 0 / 5%);
+	box-shadow:
+		inset 0 2px 4px rgb(0 0 0 / 30%),
+		0 1px 2px rgb(0 0 0 / 5%);
 }
 
 .jodit_status(
