+++ test/loader.js
@@ -1,1 +1,1 @@
-window.fileCasesJodit = ["./tests/acceptance/toolbar.test.js","./tests/acceptance/states.test.js","./tests/acceptance/security.test.js","./tests/acceptance/interface.test.js","./tests/acceptance/inline-mode.test.js","./tests/acceptance/image.test.js","./tests/acceptance/editor.test.js","./tests/acceptance/commands.test.js","../src/plugins/plugins.test.js","../src/plugins/xpath/xpath.test.js","../src/plugins/wrap-nodes/wrap-nodes.test.js","../src/plugins/table/table.test.js","../src/plugins/tab/tab.test.js","../src/plugins/symbols/symbols.test.js","../src/plugins/sticky/sticky.test.js","../src/plugins/stat/stat.test.js","../src/plugins/source/source.test.js","../src/plugins/size/size.test.js","../src/plugins/search/search.test.js","../src/plugins/resizer/resizer.test.js","../src/plugins/resize-cells/resize-cells.test.js","../src/plugins/preview/preview.test.js","../src/plugins/placeholder/placeholder.test.js","../src/plugins/paste-storage/paste-storage.test.js","../src/plugins/paste/paste.test.js","../src/plugins/ordered-list/ordered-list.test.js","../src/plugins/mobile/mobile.test.js","../src/plugins/link/link.test.js","../src/plugins/line-height/line-height.test.js","../src/plugins/limit/limit.test.js","../src/plugins/inline-popup/inline-popup.test.js","../src/plugins/indent/indent.test.js","../src/plugins/image-properties/image-properties.test.js","../src/plugins/image/image.test.js","../src/plugins/iframe/iframe.test.js","../src/plugins/hr/hr.test.js","../src/plugins/hotkeys/hotkeys.test.js","../src/plugins/fullsize/fullsize.test.js","../src/plugins/font/font.test.js","../src/plugins/focus/focus.test.js","../src/plugins/enter/enter.test.js","../src/plugins/dtd/dtd.test.js","../src/plugins/drag-and-drop-element/drag-and-drop-element.test.js","../src/plugins/delete/delete.test.js","../src/plugins/copy-format/copy-format.test.js","../src/plugins/color/color.test.js","../src/plugins/clipboard/clipboard.test.js","../src/plugins/clean-html/clean-html.test.js","../src/plugins/class-span/class-span.test.js","../src/plugins/bold/bold.test.js","../src/plugins/backspace/backspace.test.js","../src/plugins/add-new-line/add-new-line.test.js","../src/modules/uploader/uploader.test.js","../src/modules/table/table.test.js","../src/modules/messages/messages.test.js","../src/modules/history/history.test.js","../src/modules/file-browser/file-browser.test.js","../src/modules/dialog/dialog.test.js","../src/langs/i18n.test.js","../src/core/ui/popup/popup.test.js","../src/core/ui/group/group.test.js","../src/core/ui/button/tooltip/tooltip.test.js","../src/core/selection/selection.test.js","../src/core/selection/style/style.test.js","../src/core/helpers/helpers.test.js","../src/core/helpers/size/position.test.js","../src/core/event-emitter/object-observer.test.js","../src/core/event-emitter/event-emitter.test.js","../src/core/dom/dom.test.js","../src/core/decorators/decorators.test.js","../src/core/create/create.test.js","../src/core/async/async.test.js"];
+window.fileCasesJodit = ["./tests/acceptance/toolbar.test.js","./tests/acceptance/states.test.js","./tests/acceptance/security.test.js","./tests/acceptance/interface.test.js","./tests/acceptance/inline-mode.test.js","./tests/acceptance/image.test.js","./tests/acceptance/editor.test.js","./tests/acceptance/commands.test.js","../src/plugins/plugins.test.js","../src/plugins/xpath/xpath.test.js","../src/plugins/wrap-nodes/wrap-nodes.test.js","../src/plugins/table/table.test.js","../src/plugins/tab/tab.test.js","../src/plugins/symbols/symbols.test.js","../src/plugins/sticky/sticky.test.js","../src/plugins/stat/stat.test.js","../src/plugins/source/source.test.js","../src/plugins/size/size.test.js","../src/plugins/search/search.test.js","../src/plugins/resizer/resizer.test.js","../src/plugins/resize-cells/resize-cells.test.js","../src/plugins/preview/preview.test.js","../src/plugins/placeholder/placeholder.test.js","../src/plugins/paste-storage/paste-storage.test.js","../src/plugins/paste/paste.word.test.js","../src/plugins/paste/paste.test.js","../src/plugins/ordered-list/ordered-list.test.js","../src/plugins/mobile/mobile.test.js","../src/plugins/link/link.test.js","../src/plugins/line-height/line-height.test.js","../src/plugins/limit/limit.test.js","../src/plugins/inline-popup/inline-popup.test.js","../src/plugins/indent/indent.test.js","../src/plugins/image-properties/image-properties.test.js","../src/plugins/image/image.test.js","../src/plugins/iframe/iframe.test.js","../src/plugins/hr/hr.test.js","../src/plugins/hotkeys/hotkeys.test.js","../src/plugins/fullsize/fullsize.test.js","../src/plugins/font/font.test.js","../src/plugins/focus/focus.test.js","../src/plugins/enter/enter.test.js","../src/plugins/dtd/dtd.test.js","../src/plugins/drag-and-drop-element/drag-and-drop-element.test.js","../src/plugins/delete/delete.test.js","../src/plugins/copy-format/copy-format.test.js","../src/plugins/color/color.test.js","../src/plugins/clipboard/clipboard.test.js","../src/plugins/clean-html/clean-html.test.js","../src/plugins/class-span/class-span.test.js","../src/plugins/bold/bold.test.js","../src/plugins/backspace/backspace.test.js","../src/plugins/add-new-line/add-new-line.test.js","../src/modules/uploader/uploader.test.js","../src/modules/table/table.test.js","../src/modules/messages/messages.test.js","../src/modules/history/history.test.js","../src/modules/file-browser/file-browser.test.js","../src/modules/dialog/dialog.test.js","../src/langs/i18n.test.js","../src/core/ui/popup/popup.test.js","../src/core/ui/group/group.test.js","../src/core/ui/button/tooltip/tooltip.test.js","../src/core/selection/selection.test.js","../src/core/selection/style/style.test.js","../src/core/helpers/helpers.test.js","../src/core/helpers/size/position.test.js","../src/core/event-emitter/object-observer.test.js","../src/core/event-emitter/event-emitter.test.js","../src/core/dom/dom.test.js","../src/core/decorators/decorators.test.js","../src/core/create/create.test.js","../src/core/async/async.test.js"];
